repaired threading for Win11; better separation
I wasn't careful enough with the use of threads, and there were conflicts with pygame's own threads in the background. It was working on Linux, but not on Windows 11. Fixed by placing all display management (or, essentially all pygame management) within the spawned thread, while game state is kept in the original thread. This is a better separation anyway.